Reorganize result code
