+++ src/ts/main.ts
@@ -1,11 +1,11 @@
 import { initForm } from "./form/initForm";
 import { initFormState } from "./state/FormState";
-import { subscribeLinkGenerator } from "./linkGenerator";
+import initResult from "./result/initResult";
 
 export default function initApp(): void {
   const formState = initFormState();
   initForm(formState);
-  subscribeLinkGenerator(formState);
+  initResult(formState);
 
   // Debugging
   formState.subscribe((v) => console.log(v));
